+++ src/main/java/io/hhplus/tdd/point/entity/PointValidator.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.entity;
 
 import org.springframework.stereotype.Component;
