+++ src/main/java/io/hhplus/tdd/point/repository/impl/PointHistoryRepositoryImpl.java
@@ -1,8 +1,8 @@
 package io.hhplus.tdd.point.repository.impl;
 
 import io.hhplus.tdd.database.PointHistoryTable;
-import io.hhplus.tdd.point.PointHistory;
-import io.hhplus.tdd.point.TransactionType;
+import io.hhplus.tdd.point.entity.PointHistory;
+import io.hhplus.tdd.point.entity.TransactionType;
 import io.hhplus.tdd.point.repository.PointHistoryRepository;
 import org.springframework.stereotype.Repository;
