+++ src/main/java/io/hhplus/tdd/point/repository/impl/UserPointRepositoryImpl.java
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.point.repository.impl;
 
 import io.hhplus.tdd.database.UserPointTable;
-import io.hhplus.tdd.point.UserPoint;
+import io.hhplus.tdd.point.entity.UserPoint;
 import io.hhplus.tdd.point.repository.UserPointRepository;
 import org.springframework.stereotype.Repository;
